frontend/requirements.txt
gradio>=4.0.0
websockets>=12.0
httpx>=0.25.0
pillow>=10.0.0
llama-cpp-python>=0.3.0
huggingface-hub>=0.20.0
